% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dating}
\alias{dating}
\title{Dataset dating}
\format{
A data frame with 19 rows and 2 variables:
\describe{
  \item{carbon}{carbon age}
  \item{thorium}{thorium age}
  }
}
\source{
Cleveland W. S. (1993). “Visualizing Data”. Hobart Press.
}
\usage{
dating
}
\description{
From Cleveland (1993): Ages of many ancient objects are determined by carbon dating. A second dating method, first reported in 1990, provides calibration back to at least 30 kyr BP by measuring the decay of uranium to thorium. The group that invented the method took core samples in coral off the coast of Barbados and dated the material back to nearly 30 kyr BP using both the carbon and thorium methods. The thorium results were used to study the accuracy of the carbon method.
}
\keyword{datasets}
