% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etanol}
\alias{etanol}
\title{Dataset etanol}
\format{
A data frame with 88 rows and 2 variables:
\describe{
  \item{NOx}{concentration of nitric oxide plus the concentration of nitrogen dioxide normalized by the amount of work of the engine.}
  \item{C}{compression ratio to which the engine is set}
  \item{E}{equivalence ratio at which the engine was run}
  }
}
\source{
Cleveland W. S. (1993). “Visualizing Data”. Hobart Press.
}
\usage{
etanol
}
\description{
From Cleveland (1993): An experiment studied exhaust from an experimental one-cylinder engine fueled by ethanol. The response, which will be denoted by NOx, is the concentration of nitric oxide, NO, plus the concentration of nitrogen dioxide, NO2, normalized by the amount of work of the engine. The units are microg/xg of NOx per joule. One factor is the equivalence ratio, E, at which the engine was run. E is a measure of the richness of the air and fuel mixture; as E increases there is more fuel in the mixture. Another factor is C, the compression ratio to which the engine is set. C is the volume inside the cylinder when the piston is retracted, divided by the volume when the piston is at its maximum point of penetration into the cylinder. There were 88 runs of the experiment.
}
\keyword{datasets}
