% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fly}
\alias{fly}
\title{Dataset fly}
\format{
A data frame with 823 rows and 2 variables:
\describe{
  \item{facet}{number of facets of the eyes}
  \item{temperature}{incubator temperature}
  }
}
\source{
Cleveland W. S. (1993). “Visualizing Data”. Hobart Press.
}
\usage{
fly
}
\description{
From Cleveland (1993): In 1924, a journal article reported 823 observations from a genetics experiment on flies' eyes. Stocks of the ubiquitous species Drosophila melanogaster Meig were hatched in nine incubators whose temperatures varied from 15°C to 31°C in equal steps of 2°C. The number of facets of the eyes of each hatched fly were reported in units that essentially make the measurement scale logarithmic. The goal of the experiment was to see how facet number depends on temperature.
}
\keyword{datasets}
