% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{futbol}
\alias{futbol}
\title{Dataset futbol}
\format{A data frame with 300 rows and 2 variables:
\describe{
  \item{longp}{category of leg length}
  \item{dist}{kick distance}
}}
\source{
?
}
\usage{
futbol
}
\description{
Acá hay que poner la descripción del dataset
}
\keyword{datasets}
