% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tmd.R
\name{gg_tmd_paired}
\alias{gg_tmd_paired}
\title{The gg_tmd_paired function}
\usage{
gg_tmd_paired(
  df,
  vble1,
  vble2,
  xlabel = "Mean",
  ylabel = "Difference",
  loess = TRUE,
  loess_span = 1,
  loess_degree = 1,
  loess_family = "gaussian",
  ...
)
}
\arguments{
\item{df}{dataframe}

\item{vble1, vble2}{numeric variables to be analized}

\item{xlabel}{label for x-axis, defaults to "Mean"}

\item{ylabel}{label for y-axis, defaults to "Difference"}

\item{loess}{logical; should a loess smoothing curve be added to the coplots?
Defaults to TRUE.}

\item{loess_span}{span parameter for loess}

\item{loess_degree}{degree parameter for loess}

\item{loess_family}{famiyly argument for the loess() function}

\item{...}{parameters to be passed to geom_point(), such as size, color, shape.}
}
\value{
a ggplot
}
\description{
Returns Tukey's Mean-Difference plot for paired data (both variables must be measured in the same scale).
}
\details{
Differences are computed as `vble1 - vble2`.
}
\examples{
gg_tmd_paired(ozone, stamford, yonkers)
}
