% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{Dataset ozone}
\format{
A data frame with 132 rows and 2 variables:
\describe{
  \item{dia}{day}
  \item{yonkers}{air mass at Yonkers}
  \item{stamford}{air mass at Stamford}
  }
}
\source{
Cleveland W. S. (1993). “Visualizing Data”. Hobart Press.
}
\usage{
ozone
}
\description{
From Cleveland (1993): The data are daily maximum ozone concentrations at ground level on 132 days from May 1,1974 to September 30,1974 at two sites in the U.S.A. — Yonkers, New York and Stamford, Connecticut — which are approximately 30 km from one another. The sample for each measurement is the air mass on a particular day, and the bivariate data arise from two measurements at the two sites.
}
\keyword{datasets}
