% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{playfair}
\alias{playfair}
\title{Dataset playfair}
\format{
A data frame with 22 rows and 2 variables:
\describe{
  \item{city}{city}
  \item{population}{population}
  \item{diameter}{diameter of the circle in the figure}
  }
}
\source{
Cleveland W. S. (1993). “Visualizing Data”. Hobart Press.
}
\usage{
playfair
}
\description{
From Cleveland (1993): In 1801, William Playfair published his Statistical Breviary, which contains many displays of economic and demographic data. One display, beautifully reproduced by Tufte, graphs the populations of 22 cities by the areas of circles. The graph also contains a table of the populations, so we can compare the data and the areas of the circles.
}
\keyword{datasets}
