% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{polarization}
\alias{polarization}
\title{Dataset polarization}
\format{
A data frame with 355 rows and 2 variables:
\describe{
  \item{concentration}{particulate concentration}
  \item{babinet}{Babinet point}
  }
}
\source{
Cleveland W. S. (1993). “Visualizing Data”. Hobart Press.
}
\usage{
polarization
}
\description{
From Cleveland (1993): This data comes from an experiment on the scattering of sunhght in the atmosphere. One variable is the Babinet point, the scattering angle at which the polarization of sunhght vanishes. The other one is the atmospheric concentration of soHd particles in the air. The goal is to determine the dependence of the Babinet point on concentration.
}
\keyword{datasets}
