% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rubber}
\alias{rubber}
\title{Dataset rubber}
\format{A data frame with 78 rows and 2 variables:
\describe{
  \item{hardness}{...}
  \item{tensile.strength}{...}
  \item{abrasion.loss}{...}
  \item{ts.low}{...}
  \item{ts.high}{...}
}}
\source{
?
}
\usage{
rubber
}
\description{
Acá hay que poner la descripción del dataset
}
\keyword{datasets}
