% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_pt.R
\name{transf_pot}
\alias{transf_pot}
\title{The trasf_pot function}
\usage{
transf_pot(x, tau = 0)
}
\arguments{
\item{x}{numeric vector to be transformed}

\item{tau}{powers}
}
\value{
vector of transformed values
}
\description{
Helper function for gg_pt
}
\keyword{internal}
