% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fusion}
\alias{fusion}
\title{Dataset fusion}
\format{A data frame with 78 rows and 2 variables:
\describe{
  \item{time}{tiempo de fusión, segundos}
  \item{nv.vv}{grupo}
}}
\source{
?
}
\usage{
fusion
}
\description{
Acá hay que poner la descripción del dataset
}
\keyword{datasets}
