% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coplots.R
\name{gg_coplot}
\alias{gg_coplot}
\title{The gg_coplot function}
\usage{
gg_coplot(
  df,
  x,
  y,
  faceting,
  number_bins = 6,
  overlap = 0.5,
  loess = TRUE,
  loess_span = 3/4,
  loess_degree = 1,
  loess_family = "gaussian",
  ylabel = quo_text(y),
  xlabel = quo_text(x),
  facetlabel = quo_text(faceting),
  show_intervals = TRUE,
  intervals_height = 0.25,
  remove_strip = FALSE,
  facets_nrow = NULL,
  hline_at = NULL,
  ...
)
}
\arguments{
\item{df}{dataframe}

\item{x}{numeric variable for x-axis}

\item{y}{numeric variable for y-axis}

\item{faceting}{faceting numeric variable}

\item{number_bins}{integer; the number of conditioning intervals}

\item{overlap}{numeric < 1; the fraction of overlap of the conditioning
variables}

\item{loess}{logical; should a loess smoothing curve be added to the coplots?
Defaults to TRUE.}

\item{loess_span}{span parameter for loess}

\item{loess_degree}{degree parameter for loess}

\item{loess_family}{famiyly argument for the loess() function}

\item{ylabel}{label for y-axis}

\item{xlabel}{label for x-axis}

\item{facetlabel}{label for faceting variable}

\item{show_intervals}{logical; should the overlapping intervals be shown on
their own panel on the top of the figure? Defaults to TRUE.}

\item{intervals_height}{numeric between 0 and 1, relative size of the
intervals pane}

\item{remove_strip}{logical; should de facets have no strips with labels?
Default to FALSE.}

\item{facets_nrow}{integer; number of rows for the facets}

\item{hline_at}{numeric; if provide a horizontal line will be added at that
heigth}

\item{...}{addtional parameters passed to geom_point()}
}
\value{
a coplot
}
\description{
Implements conditional plots or coplots.
}
\examples{
data(rubber)
gg_coplot(rubber, x = tensile.strength, y = abrasion.loss, faceting = hardness,
  number_bins = 6, overlap = 3/4,
  ylabel = "Pérdida de abrasión (g/hp-hour))",
  xlabel = "Resistencia a la tracción (kg/cm2)",
  facetlabel = "Dureza (grados Shore)", loess_family = "symmetric", size = 2)

}
