% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coplots.R
\name{make_coplot_df}
\alias{make_coplot_df}
\title{The make_coplot_df function}
\usage{
make_coplot_df(df, vble, number_bins = 6, overlap = 0.5)
}
\arguments{
\item{df}{dataframe}

\item{vble}{faceting numeric variable}

\item{number_bins}{integer; the number of conditioning intervals}

\item{overlap}{numeric < 1; the fraction of overlap of the conditioning variables}
}
\value{
a dataset to be used in the creation of coplots
}
\description{
It creates dataframes to be used in coplot
}
\details{
Adapted from \href{https://jfukuyama.github.io/teaching/stat670/notes/lecture-11.html}{here}.
}
\examples{
data(rubber)
data_coplot <- make_coplot_df(rubber, hardness, 6, 3/4)
}
